+++ src/in_out.rs
@@ -18,6 +18,8 @@
 // *************************************************************************
 
 use std::any::Any;
+#[cfg(feature = "readline")]
+use std::ffi::CString;
 
 use gui::Cap;
 use gui::Event;
@@ -28,6 +30,9 @@ use gui::OptionChain;
 use gui::UiEvent;
 use gui::UiEvents;
 use gui::Widget;
+
+#[cfg(feature = "readline")]
+use rline::Readline;
 
 use crate::event::EventUpdate;
 use crate::termui::TermUiEvent;
@@ -41,6 +46,18 @@ pub enum InOut {
   Error(String),
   Input(String, usize),
   Clear,
+}
+
+#[cfg(feature = "readline")]
+impl InOut {
+  /// Check whether the `InOut` state is `Input`.
+  fn is_input(&self) -> bool {
+    if let InOut::Input(..) = &self {
+      return true
+    } else {
+      return false
+    }
+  }
 }
 
 
@@ -87,6 +104,8 @@ pub struct InOutArea {
   id: Id,
   prev_focused: Option<Id>,
   in_out: InOutState,
+  #[cfg(feature = "readline")]
+  readline: Readline,
 }
 
 impl InOutArea {
@@ -100,6 +119,8 @@ impl InOutArea {
       id: id,
       prev_focused: None,
       in_out: Default::default(),
+      #[cfg(feature = "readline")]
+      readline: Readline::new(),
     }
   }
 
@@ -112,6 +133,18 @@ impl InOutArea {
     self.in_out.bump();
 
     if in_out != *self.in_out.get() {
+      #[cfg(feature = "readline")]
+      {
+        if let InOut::Input(s, idx) = &in_out {
+          // We clear the undo buffer if we transition from a non-Input
+          // state to an Input state. Input-to-Input transitions are
+          // believed to be those just updating the text the user is
+          // working on already.
+          let cstr = CString::new(s.clone()).unwrap();
+          let clear_undo = !self.in_out.get().is_input();
+          self.readline.reset(cstr, *idx, clear_undo);
+        }
+      }
       self.in_out.set(in_out);
       (None as Option<Event>).update()
     } else {
@@ -174,7 +207,7 @@ impl InOutArea {
           None
         }
       },
-      #[cfg(test)]
+      #[cfg(all(test, not(feature = "readline")))]
       TermUiEvent::GetInOut => {
         let resp = TermUiEvent::GetInOutResp(self.in_out.get().clone());
         Some(UiEvent::Custom(Box::new(resp)).into())
@@ -183,7 +216,21 @@ impl InOutArea {
     }
   }
 
+  /// Finish text input by changing the internal state and emitting an event.
+  fn finish_input(&mut self, string: Option<String>, cap: &mut dyn Cap) -> Option<UiEvents> {
+    let update = self.change_state(InOut::Clear);
+    let widget = self.restore_focus(cap);
+    let event = if let Some(s) = string {
+      Box::new(TermUiEvent::EnteredText(s))
+    } else {
+      Box::new(TermUiEvent::InputCanceled)
+    };
+    debug_assert!(update.is_some());
+    Some(UiEvent::Directed(widget, event)).chain(update)
+  }
+
   /// Handle a key press.
+  #[cfg(not(feature = "readline"))]
   fn handle_key(&mut self,
                 mut s: String,
                 mut idx: usize,
@@ -192,15 +239,8 @@ impl InOutArea {
     match key {
       Key::Esc |
       Key::Return => {
-        let update = self.change_state(InOut::Clear);
-        let widget = self.restore_focus(cap);
-        let event = if key == Key::Return {
-          Box::new(TermUiEvent::EnteredText(s))
-        } else {
-          Box::new(TermUiEvent::InputCanceled)
-        };
-        debug_assert!(update.is_some());
-        Some(UiEvent::Directed(widget, event)).chain(update)
+        let string = if key == Key::Return { Some(s) } else { None };
+        self.finish_input(string, cap)
       },
       Key::Char(c) => {
         s.insert(idx, c);
@@ -252,6 +292,72 @@ impl InOutArea {
         }
       },
       _ => None,
+    }
+  }
+
+  /// Handle a key press.
+  #[cfg(feature = "readline")]
+  fn handle_key(&mut self,
+                _s: String,
+                idx: usize,
+                key: Key,
+                cap: &mut dyn Cap) -> Option<UiEvents> {
+    const BACKSPACE: char = 0x08 as char;
+    const DELETE: char = 0x7f as char;
+    const ESCAPE: char = 0x1b as char;
+
+    // We only support a couple of very basic keys when running in
+    // readline enabled mode. The entire idea behind using libreadline
+    // is that we can stay closer to the central keyboard position and
+    // don't have to use, say, the arrow keys. Doing more would also
+    // require us to map more keys to much more complex character
+    // sequences that libreadline understands which in turn basically
+    // means encoding deep terminal knowledge which we don't want to
+    // have to deal with.
+    let c = match key {
+      Key::Return => Some('\n'),
+      Key::Backspace => Some(BACKSPACE),
+      Key::Delete => Some(DELETE),
+      Key::Esc => Some(ESCAPE),
+      Key::Char(c) => Some(c),
+      _ => None,
+    };
+
+    if let Some(c) = c {
+      match self.readline.feed(c as i32) {
+        Some(line) => self.finish_input(Some(line.into_string().unwrap()), cap),
+        None => {
+          let (s_, idx_) = self.readline.inspect(|s, pos| (s.to_owned(), pos));
+          // We treat Esc a little specially. In a vi-mode enabled
+          // configuration of libreadline Esc cancels input mode when we
+          // are in it, and does nothing otherwise. That is what we are
+          // interested in here. So we peek at the index we get and see
+          // if it changed (because leaving input mode moves the cursor
+          // to the left by one). If nothing changed, then we actually
+          // cancel the text input. That is not the nicest logic, but
+          // the only way we have found that accomplishes what we want.
+          if key == Key::Esc && idx_ == idx {
+            // TODO: We have a problem here. What may end up happening
+            //       is that we disrupt libreadline's workflow by
+            //       effectively canceling what it was doing. If, for
+            //       instance, we were in vi-movement-mode and we simply
+            //       stop the input process libreadline does not know
+            //       about that and will stay in this mode. So next time
+            //       we start editing again, we will still be in this
+            //       mode. Unfortunately, rline's reset does not deal
+            //       with this case (perhaps rightly so). For now, just
+            //       create a new `Readline` context and that will take
+            //       care of resetting things to the default (which is
+            //       input mode).
+            self.readline = Readline::new();
+            self.finish_input(None, cap)
+          } else {
+            self.change_state(InOut::Input(s_.into_string().unwrap(), idx_))
+          }
+        },
+      }
+    } else {
+      None
     }
   }